requirements.txt
tqdm
yt-dlp
pydub
openai-whisper
pandas